+++ core/src/utils/mod.rs
@@ -1,4 +1,3 @@
 pub mod html;
 pub mod spinner;
-#[cfg(test)]
 pub mod test_utils;